messed up merge a bit, partial fix
