+++ client/src/routes/EditorPage/CardNavigation.jsx
