+++ client/src/routes/EditorPage/Logo.jsx
