+++ client/src/routes/EditorPage/SearchBar.jsx
