+++ client/src/routes/LoginPage/LoginForm.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+// import { Link } from 'react-router-dom'
 
 export const LoginForm = (props) => {
   const { email, setEmail } = useState('')
@@ -9,41 +10,20 @@ export const LoginForm = (props) => {
     console.log(email)
   }
 
-  async function loginUser(event) {
-    event.preventDefault();
-    const username = document.getElementById('username').value;
-    const password = document.getElementById('password').value;
-
-    const result = await fetch('http://localhost:9999/api/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username,
-        password
-      })
-    }).then((res) => res.json())
-
-    if(result.status === 'ok'){
-      console.log("Got the token: ", result.data)
-    } else{
-      alert(result.error)
-    }
-    ;
-  }
-
   return (
     <div className='auth-form-container'>
       <h2>Login</h2>
-      <form className='login-form' onSubmit={loginUser}>
-        <label htmlFor='username'>username</label>
-        <input type='text' placeholder='username' id='username' />
+      <form className='login-form' onSubmit={handleSubmit}>
+        <label htmlFor='email'>email</label>
+        <input value={email} onChange={(e) => setEmail(e.target.value)} type='email' placeholder='youremail@gmail.com' id='email' name='email' />
         <label htmlFor='password'>password</label>
-        <input type='password' placeholder='********' id='password' name='password' />
+        <input value={pass} onChange={(e) => setPass(e.target.value)} type='password' placeholder='********' id='password' name='password' />
         <button className='login-button'>Login</button>
       </form>
       <button className='link-btn' onClick={() => props.onFormSwitch('register')}>Don't have an account?</button>
+      {/* <Link to='/EditorPage' className='nav-links'>
+          Demo
+      </Link> */}
     </div>
   )
 }
