+++ client/src/routes/LoginPage/LoginPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 // import logo from './logo.svg';
-import '../../App.css'
+import './LoginPage.css'
 import { LoginForm } from './LoginForm.jsx'
 import { RegisterForm } from './RegisterForm.jsx'
