+++ client/src/routes/LoginPage/RegisterForm.jsx
@@ -1,46 +1,28 @@
-import React from 'react';
+import React, { useState } from 'react'
 
 export const RegisterForm = (props) => {
-  async function registerUser(event) {
-    event.preventDefault();
-    const username = document.getElementById('name').value;
-    const email = document.getElementById('email').value;
-    const password = document.getElementById('password').value;
+  const { email, setEmail } = useState('')
+  const { pass, setPass } = useState('')
+  const { name, setName } = useState('')
 
-    const result = await fetch('http://localhost:9999/api/register', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username,
-        email,
-        password
-      })
-    }).then((res) => res.json())
-
-    if(result.status === 'ok'){
-
-    } else{
-      alert(result.error)
-    }
-
-    ;
+  const handleSubmit = (e) => {
+    e.preventDefault()
+    console.log(email)
   }
 
   return (
     <div className='auth-form-container'>
       <h2>Register</h2>
-      <form className='register-form' onSubmit={registerUser}>
+      <form className='register-form' onSubmit={handleSubmit}>
         <label htmlFor='name'>name</label>
-        <input type='text' placeholder='name' id='name' name='name' />
+        <input value={name} onChange={(e) => setName(e.target.value)} type='name' placeholder='name' id='name' name='name' />
         <label htmlFor='email'>email</label>
-        <input type='email' placeholder='youremail@gmail.com' id='email' name='email' />
+        <input value={email} onChange={(e) => setEmail(e.target.value)} type='email' placeholder='youremail@gmail.com' id='email' name='email' />
         <label htmlFor='password'>password</label>
-        <input type='password' placeholder='********' id='password' name='password' />
-        <input type="submit" value="Submit Form" />
+        <input value={pass} onChange={(e) => setPass(e.target.value)} type='password' placeholder='********' id='password' name='password' />
+        <button>Register</button>
       </form>
       <button className='link-btn' onClick={() => props.onFormSwitch('login')}>Already have an account?</button>
     </div>
-  );
-};
+  )
+}
