--- client/src/routes/MainCardDisplay/Card.js
@@ -1,13 +1,0 @@
-import React from 'react'
-// import { useState } from 'react'
-// import logo from './logo.svg'
-import { CardPage } from './CardPage.jsx'
-import './CardPage.css'
-
-function Card () {
-  return (
-    <CardPage> </CardPage>
-  )
-}
-
-export default Card
