+++ package.json
@@ -4,18 +4,19 @@
     "description": "",
     "main": "",
     "scripts": {
-      "test": "standard --fix && react-scripts test"
+        "test": "standard --fix && react-scripts test"
     },
     "keywords": [],
     "author": "",
     "license": "ISC",
     "dependencies": {
-      "cors": "^2.8.5",
-      "dotenv": "^16.1.4",
-      "express": "^4.18.2",
-      "mongodb": "^5.6.0"
+        "cors": "^2.8.5",
+        "dotenv": "^16.1.4",
+        "express": "^4.18.2",
+        "mongodb": "^5.6.0",
+        "react-router-dom": "^6.14.2"
     },
     "devDependencies": {
-      "standard": "*"
+        "standard": "*"
     }
 }
